Move rate limiter tests to avoid make an internal class public
